Language.Dash.Parser

Documentation

lambda :: DashParser (Term String)

Parses a lambda expression. A lambda expression is given by the unicode character λ followed immediately by a variable name or names, then a . followed by the body of the lambda function.

If more than one variable appears after the λ, we treat this as syntactic sugar. For example:

>>> (λx y z. foo)

...is parsed the same as...

>>> (λx. (λy. (λz. foo)))

variable :: DashParser (Term String)

Parses a variable. A variable in dash is given by $ followed by an alphanumeric identifier, because we like stealing crappy ideas from PHP, apparently.

expression :: DashParser (Term String)

An expression is where the magic happens. It is a formulation of all of the other parsers we define. It is comprised of literals (such as literalInt and literalString) and s-expressions ((...) lambdas, if-then-else expressions, letrec bindings, etc.)

expressions :: DashParser [Term String]

Generates a list of expressions up to eof and returns them.

letRecBinding :: DashParser (Term String)

Parses a letrec-binding-based expression. This is given by the keyword letrec followed by a list of bindings enclosed in [...], followed by the body of the expression.

Bindings have the form: 'variable'='expression'. They are separated by spaces.

If more than one binding appears in the list, we treat this as syntactic sugar. For example:

>>> (letrec [$foo=1 $bar=2 $baz=$foo] $baz)

...is parsed the same as...

>>> (letrec [$foo=1] (letrec [$bar=2] (letrec [$baz=$foo] $baz)))

literalBool :: DashParser (Term String)

Parses a literal Bool, given by the keywords true or false.

literalInt :: DashParser (Term String)

Parses a literal Integer.

literalString :: DashParser (Term String)

Parses a literal String, which is enclosed in double quotes.

runParser :: DashParser a -> Parser a